rename customlogger to logger
